% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot Gene Expression Distribution}
\usage{
plot_expression(data, gene, method = c("base", "lattice", "ggplot2"))
}
\arguments{
\item{data}{An object of class `scData`.}

\item{gene}{A character string specifying the gene.}

\item{method}{Plotting system to use.}
}
\value{
A plot object or NULL.
}
\description{
Generates boxplots or violins depicting the statistical distribution (spread and median)
  of transcript counts for a specific targeting gene across different clusters.
}
\examples{
data <- load_scrna()
plot_expression(data, "CD4", method = "lattice")
}
