% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA Emdeddings}
\usage{
plot_pca(data, method = c("base", "lattice", "ggplot2"))
}
\arguments{
\item{data}{An object of class `scData`.}

\item{method}{Plotting system to use: "base", "lattice", or "ggplot2".}
}
\value{
A plot object or NULL.
}
\description{
Maps the highly complex sequencing dimensions down to just 2 Principal Components (PC1 and PC2).
  Cells with similar transcriptomic profiles (similar biological identities) map closer together.
}
\examples{
data <- load_scrna()
plot_pca(data, method = "ggplot2")
}
