% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scData.R
\name{print.scData}
\alias{print.scData}
\title{Print method for scData}
\usage{
\method{print}{scData}(x, ...)
}
\arguments{
\item{x}{An object of class `scData`}

\item{...}{Additional arguments}
}
\description{
Briefly summarizes the matrix dimensions.
}
