% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scData.R
\name{summary.scData}
\alias{summary.scData}
\title{Summary method for scData}
\usage{
\method{summary}{scData}(object, ...)
}
\arguments{
\item{object}{An object of class `scData`}

\item{...}{Additional arguments}
}
\description{
Computes library sizes (total gene counts per cell) and population counts.
}
